fix: change modal Buy a library card
