+++ library/index.html
@@ -146,7 +146,7 @@
             </ul>
           </div>
           <!-- <List books> -->
-          <div class="favorites__books">
+          <div class="favorites__books" id="favorites__books">
             <div class="favorites__book">
               <div class="favorites__book-title">Staff Picks</div>
               <div class="favorites__book-name">The Book Eaters</div>
@@ -655,7 +655,7 @@
         </form>
         <div class="modal-register__register">
           <p class="modal__question">Already have an account?</p>
-          <button class="modal-register__button-register">Login</button>
+          <button class="modal-register__button-login">Login</button>
         </div>
         <button class="modal__button-close"></button>
       </div>
